chore: Add support for `_snoop` module in `ll/__init__.py`
This commit adds support for the `_snoop` module in the `ll/__init__.py` file. The `_snoop` module provides a `snoop` function that can be used for debugging and inspecting code execution. This feature enhances the debugging capabilities of the `ll` module by allowing developers to easily trace and analyze the execution flow.